api/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.1
mysql-connector-python==9.0.0
python-dotenv==1.0.1
requests==2.32.3